Move all index related consts into index_param.h (#560)
