--- include/knowhere/comp/metric_type.h
@@ -1,26 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-// -*- c++ -*-
-
-#ifndef KNOWHERE_METRIC_TYPE_H
-#define KNOWHERE_METRIC_TYPE_H
-
-namespace knowhere {
-namespace metric {
-constexpr const char* IP = "IP";
-constexpr const char* L2 = "L2";
-constexpr const char* HAMMING = "HAMMING";
-constexpr const char* JACCARD = "JACCARD";
-constexpr const char* TANIMOTO = "TANIMOTO";
-constexpr const char* SUBSTRUCTURE = "SUBSTRUCTURE";
-constexpr const char* SUPERSTRUCTURE = "SUPERSTRUCTURE";
-}  // namespace metric
-
-}  // namespace knowhere
-
-#endif
