+++ src/index/hnsw/hnsw.cc
@@ -7,7 +7,7 @@
 #include "hnswlib/hnswalg.h"
 #include "hnswlib/hnswlib.h"
 #include "index/hnsw/hnsw_config.h"
-#include "knowhere/comp/metric_type.h"
+#include "knowhere/comp/index_param.h"
 #include "knowhere/knowhere.h"
 
 namespace knowhere {
